LayoutTests/http/tests/security/shape-image-cors-redirect-error-message-logging-4-expected.txt

CONSOLE MESSAGE: Cross-origin redirection to mailto://example.com denied by Cross-Origin Resource Sharing policy: URL is either a non-HTTP URL or contains credentials.
Verify the error message in console in case of CORS failing checks.
